Revert change to Viewport, adding a RTL provider component
