+++ packages/common/src/styles/RTLProvider.tsx
@@ -1,0 +1,8 @@
+import React, { FC } from 'react';
+import { useRtl } from '../intl/intlHelpers';
+
+export const RTLProvider: FC = props => {
+  const isRtl = useRtl();
+
+  return <div style={{ direction: isRtl ? 'rtl' : 'ltr' }} {...props} />;
+};
