+++ packages/common/src/styles/index.ts
@@ -1,4 +1,4 @@
-import { AppGlobalStyles } from './AppGlobalStyles';
+import { RTLProvider } from './RTLProvider';
 import { useAppTheme } from './useAppTheme';
 import AppThemeProvider from './ThemeProvider';
 import { styled } from '@material-ui/core/styles';
@@ -15,5 +15,5 @@ export {
   useMediaQuery,
   useTheme,
   useAppTheme,
-  AppGlobalStyles,
+  RTLProvider,
 };
