+++ packages/common/src/ui/components/index.ts
@@ -2,6 +2,7 @@ import Button from '@material-ui/core/Button';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import Collapse from '@material-ui/core/Collapse';
 import Divider from '@material-ui/core/Divider';
+import GlobalStyles from '@material-ui/core/GlobalStyles';
 import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
 import Typography from '@material-ui/core/Typography';
@@ -10,6 +11,7 @@ export * from './LoadingApp';
 export * from './RemoteComponent';
 export { AppNavLink } from './NavLink';
 export {
+  GlobalStyles,
   Button,
   CircularProgress,
   Collapse,
