+++ packages/host/src/Viewport.tsx
@@ -1,9 +1,28 @@
 import React from 'react';
-import { CssBaseline } from '@openmsupply-client/common';
+import { GlobalStyles, CssBaseline } from '@openmsupply-client/common';
 
 const Viewport: React.FC = props => {
   return (
     <React.Fragment>
+      <GlobalStyles
+        styles={{
+          '*:-webkit-full-screen': {
+            height: '100%',
+            width: '100%',
+          },
+          '#root': {
+            height: '100%',
+            display: 'flex',
+            flexDirection: 'column',
+          },
+          html: { position: 'fixed' },
+          'html, body': {
+            height: '100%',
+            width: '100%',
+          },
+        }}
+        {...props}
+      />
       <CssBaseline />
       {props.children}
     </React.Fragment>
